leaks

+++ libft/libft.h
@@ -22,7 +22,7 @@ int				ft_strlen(char *str);
 int				ft_strcmp(char *s1, char *s2);
 int				ft_isalpha(int c);
 int				ft_isalnum(int c);
-int				ft_atoi(char *str);
+long			ft_atoi(char *str);
 int				ft_isdigit(int c);
 
 #endif
